step5まで完了
